requirements.txt
# Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
# SPDX-License-Identifier: MIT-0

aws-cdk-lib==2.65.0
constructs==10.1.253
python-dateutil~=2.8.2

amazon-braket-sdk
cdk-nag
bandit
pip-audit